Remove typography definition in theme and fix some type issues in Ledger ( need to be refactored)
