+++ src/theme/theme.ts
@@ -25,20 +25,4 @@ export const themeOptions: ThemeOptions = {
       500: '#7B8298',
     },
   },
-  typography: {
-    fontFamily: "Montserrat, sans-serif",
-    fontSize: 24,
-    h3: {
-      fontFamily: "Montserrat, sans-serif",
-      fontWeight: 700,
-      whiteSpace: "pre-line",
-      fontSize: "3rem",
-    },
-    h4: {
-      fontFamily: "Montserrat, sans-serif",
-      fontSize: "1.5rem",
-      fontWeight: 500,
-      whiteSpace: "pre-line",
-    }
-  },
 };
